+++ src/Models/Traits/TrailModelTrait.php
@@ -7,9 +7,7 @@ namespace Hirtz\Skeleton\Models\Traits;
 use Hirtz\Skeleton\Behaviors\TrailBehavior;
 use Hirtz\Skeleton\Db\ActiveRecord;
 use Hirtz\Skeleton\Models\Collections\TrailModelCollection;
-use Hirtz\Skeleton\Models\Interfaces\AdminRouteInterface;
 use Hirtz\Skeleton\Models\Interfaces\CustomAttributeInterface;
-use ReflectionClass;
 
 trait TrailModelTrait
 {
@@ -38,21 +36,6 @@ trait TrailModelTrait
         /** @var TrailBehavior $behavior */
         $behavior = $this->getBehavior('TrailBehavior');
         return $behavior;
-    }
-
-    public function getTrailModelAdminRoute(): array|false
-    {
-        return $this instanceof AdminRouteInterface ? $this->getAdminRoute() : false;
-    }
-
-    public function getTrailModelName(): string
-    {
-        return (new ReflectionClass(static::class))->getShortName();
-    }
-
-    public function getTrailModelType(): ?string
-    {
-        return null;
     }
 
     public function getTrailParents(): ?array