+++ src/Models/User.php
@@ -10,12 +10,12 @@ use davidhirtz\yii2\datetime\DateTimeBehavior;
 use Hirtz\Skeleton\Behaviors\TimestampBehavior;
 use Hirtz\Skeleton\Behaviors\TrailBehavior;
 use Hirtz\Skeleton\Db\ActiveRecord;
-use Hirtz\Skeleton\Models\Interfaces\AdminRouteInterface;
 use Hirtz\Skeleton\Models\Interfaces\CustomAttributeInterface;
 use Hirtz\Skeleton\Models\Interfaces\SearchableInterface;
 use Hirtz\Skeleton\Models\Interfaces\StatusAttributeInterface;
 use Hirtz\Skeleton\Models\Interfaces\TrailModelInterface;
 use Hirtz\Skeleton\Models\Queries\UserQuery;
+use Hirtz\Skeleton\Models\Traits\AdminModelTrait;
 use Hirtz\Skeleton\Models\Traits\CustomAttributesTrait;
 use Hirtz\Skeleton\Models\Traits\SearchableTrait;
 use Hirtz\Skeleton\Models\Traits\StatusAttributeTrait;
@@ -53,8 +53,9 @@ use yii\web\IdentityInterface;
  *
  * @mixin TrailBehavior
  */
-class User extends ActiveRecord implements AdminRouteInterface, CustomAttributeInterface, IdentityInterface, SearchableInterface, StatusAttributeInterface, TrailModelInterface
+class User extends ActiveRecord implements CustomAttributeInterface, IdentityInterface, SearchableInterface, StatusAttributeInterface, TrailModelInterface
 {
+    use AdminModelTrait;
     use CustomAttributesTrait;
     use SearchableTrait;
     use StatusAttributeTrait;
@@ -390,12 +391,12 @@ class User extends ActiveRecord implements AdminRouteInterface, CustomAttributeI
         ]);
     }
 
-    public function getTrailModelName(): string
+    public function getAdminName(): string
     {
-        return $this->id ? $this->getUsername() : $this->getTrailModelType();
+        return $this->id ? $this->getUsername() : $this->getAdminType();
     }
 
-    public function getTrailModelType(): string
+    public function getAdminType(): string
     {
         return Yii::t('skeleton', 'COMMON_USER');
     }